requirements_optimized.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
pydantic==2.5.0
PyYAML==6.0.1
httpx==0.25.2
sentence-transformers==2.2.2
torch==2.1.0
scikit-learn==1.3.2
numpy==1.24.4
PyMuPDF==1.23.8
requests==2.31.0
aiofiles==23.2.0